core(lantern): refactor Trace Engine specific code to file
