+++ core/computed/metrics/lantern-metric.js
@@ -10,7 +10,7 @@ import {LoadSimulator} from '../load-simulator.js';
 import {ProcessedNavigation} from '../processed-navigation.js';
 import {PageDependencyGraph} from '../page-dependency-graph.js';
 import {TraceEngineResult} from '../trace-engine-result.js';
-import {createProcessedNavigation} from '../../lib/lantern/lantern.js';
+import {createProcessedNavigation} from '../../lib/lantern/trace-engine-computation-data.js';
 
 /**
  * @param {LH.Artifacts.MetricComputationDataInput} data
