+++ core/lib/lantern/lantern.js
@@ -4,11 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import * as Lantern from './types/lantern.js';
-
-/** @typedef {import('@paulirish/trace_engine/models/trace/handlers/PageLoadMetricsHandler.js').MetricName} MetricName */
-/** @typedef {import('@paulirish/trace_engine/models/trace/handlers/PageLoadMetricsHandler.js').MetricScore} MetricScore */
-
 /** @type {LH.Util.SelfMap<LH.Crdp.Network.ResourceType>} */
 const NetworkRequestTypes = {
   XHR: 'XHR',
@@ -31,48 +26,6 @@ const NetworkRequestTypes = {
   Prefetch: 'Prefetch',
 };
 
-/**
- * @param {LH.Artifacts.TraceEngineResult} traceEngineResult
- * @return {Lantern.Simulation.ProcessedNavigation}
- */
-function createProcessedNavigation(traceEngineResult) {
-  const Meta = traceEngineResult.data.Meta;
-  const frameId = Meta.mainFrameId;
-  const scoresByNav = traceEngineResult.data.PageLoadMetrics.metricScoresByFrameId.get(frameId);
-  if (!scoresByNav) {
-    throw new Error('missing metric scores for main frame');
-  }
-
-  const lastNavigationId = Meta.mainFrameNavigations.at(-1)?.args.data?.navigationId;
-  const scores = lastNavigationId && scoresByNav.get(lastNavigationId);
-  if (!scores) {
-    throw new Error('missing metric scores for specified navigation');
-  }
-
-  /** @param {MetricName} metric */
-  const getTimestampOrUndefined = metric => {
-    const metricScore = scores.get(metric);
-    if (!metricScore?.event) return;
-    return metricScore.event.ts;
-  };
-  /** @param {MetricName} metric */
-  const getTimestamp = metric => {
-    const metricScore = scores.get(metric);
-    if (!metricScore?.event) throw new Error(`missing metric: ${metric}`);
-    return metricScore.event.ts;
-  };
-  // TODO: should use `MetricName.LCP`, but it is a const enum.
-  const FCP = /** @type {MetricName} */('FCP');
-  const LCP = /** @type {MetricName} */('LCP');
-  return {
-    timestamps: {
-      firstContentfulPaint: getTimestamp(FCP),
-      largestContentfulPaint: getTimestampOrUndefined(LCP),
-    },
-  };
-}
-
 export {
   NetworkRequestTypes,
-  createProcessedNavigation,
 };
